# docs/swagger.yaml
basePath: /api
definitions:
  httpErrors.ErrResponse:
    properties:
      msg:
        description: application-level error message, for debugging
        type: string
      status:
        description: http response status code
        type: integer
      statusText:
        description: user-level status message
        type: string
    type: object
...
  presenter.UserCreate:
    properties:
      confirm_password:
        example: password
        minLength: 8
        type: string
      email:
        example: hiennguyen9874@gmail.com
        type: string
      name:
        example: Xuan Hien
        type: string
      password:
        example: password
        minLength: 8
        type: string
    required:
    - confirm_password
    - email
    - name
    - password
    type: object
  presenter.UserUpdate:
    properties:
      name:
        example: Xuan Hien
        type: string
    type: object
  presenter.UserUpdatePassword:
    properties:
      confirm_password:
        example: password
        minLength: 8
        type: string
      new_password:
        example: password
        minLength: 8
        type: string
      old_password:
        example: old_password
        minLength: 8
        type: string
    required:
    - confirm_password
    - new_password
    - old_password
    type: object
  responses.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/httpErrors.ErrResponse'
      is_success:
        type: boolean
    type: object
info:
  contact: {}
  title: Go boilerplate
  version: "1.0"
paths:
...
  /user:
    get:
      consumes:
      - application/json
      description: Retrieve users.
      parameters:
      - description: limit
        format: limit
        in: query
        name: limit
        type: integer
      - description: offset
        format: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - OAuth2Password: []
      summary: Read Users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create new user.
      parameters:
      - description: Add user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/presenter.UserCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - OAuth2Password: []
      summary: Create User
      tags:
      - users
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an user by ID.
      parameters:
      - description: User Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - OAuth2Password: []
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get user by ID.
      parameters:
      - description: User Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - OAuth2Password: []
      summary: Read user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an user by ID.
      parameters:
      - description: User Id
        in: path
        name: id
        required: true
        type: string
      - description: Update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/presenter.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - OAuth2Password: []
      summary: Update user
      tags:
      - users
  /user/{id}/logoutall:
    get:
      consumes:
      - application/json
      description: Logout all session of user with id.
      parameters:
      - description: User Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - OAuth2Password: []
      summary: Logout all of user
      tags:
      - users
  /user/{id}/updatepass:
    patch:
      consumes:
      - application/json
      description: Update password user by ID.
      parameters:
      - description: User Id
        in: path
        name: id
        required: true
        type: string
      - description: Update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/presenter.UserUpdatePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - OAuth2Password: []
      summary: Update password user
      tags:
      - users
  /user/me:
    get:
      consumes:
      - application/json
      description: Get user me.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - OAuth2Password: []
      summary: Read user me
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user me.
      parameters:
      - description: Update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/presenter.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - OAuth2Password: []
      summary: Update user me
      tags:
      - users
  /user/me/updatepass:
    patch:
      consumes:
      - application/json
      description: Update password user me.
      parameters:
      - description: Update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/presenter.UserUpdatePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - OAuth2Password: []
      summary: Update password user me
      tags:
      - users
securityDefinitions:
  OAuth2Password:
    flow: password
    tokenUrl: /api/auth/login
    type: oauth2
swagger: "2.0"